IOS-8356 Update Redirecting module
