+++ Tangem/Modules/Send/Factories/StepBuilders/OnrampRedirectingBuilder.swift
@@ -9,14 +9,6 @@
 import Foundation
 import TangemExpress
 
-protocol OnrampRedirectingInput: AnyObject {
-    var selectedOnrampProvider: OnrampProvider? { get }
-}
-
-protocol OnrampRedirectingOutput: AnyObject {
-    func redirectDataDidLoad(data: OnrampRedirectData)
-}
-
 struct OnrampRedirectingBuilder {
     typealias IO = (input: OnrampRedirectingInput, output: OnrampRedirectingOutput)
     typealias ReturnValue = OnrampRedirectingViewModel
@@ -48,45 +40,5 @@ private extension OnrampRedirectingBuilder {
             output: io.output,
             onrampManager: onrampManager
         )
-    }
-}
-
-protocol OnrampRedirectingInteractor {
-    var onrampProvider: OnrampProvider? { get }
-
-    func loadRedirectData() async throws
-}
-
-class CommonOnrampRedirectingInteractor {
-    private weak var input: OnrampRedirectingInput?
-    private weak var output: OnrampRedirectingOutput?
-
-    private let onrampManager: OnrampManager
-
-    init(
-        input: OnrampRedirectingInput,
-        output: OnrampRedirectingOutput,
-        onrampManager: OnrampManager
-    ) {
-        self.input = input
-        self.output = output
-        self.onrampManager = onrampManager
-    }
-}
-
-// MARK: - OnrampRedirectingInteractor
-
-extension CommonOnrampRedirectingInteractor: OnrampRedirectingInteractor {
-    var onrampProvider: TangemExpress.OnrampProvider? {
-        input?.selectedOnrampProvider
-    }
-
-    func loadRedirectData() async throws {
-        guard let provider = input?.selectedOnrampProvider else {
-            throw CommonError.noData
-        }
-
-        let redirectData = try await onrampManager.loadRedirectData(provider: provider)
-        output?.redirectDataDidLoad(data: redirectData)
     }
 }
