+++ Tangem/Modules/Send/UI/Onramp/OnrampStep.swift
@@ -24,6 +24,10 @@ class OnrampStep {
         self.viewModel = viewModel
         self.interactor = interactor
     }
+
+    func set(router: OnrampSummaryRoutable) {
+        viewModel.router = router
+    }
 }
 
 // MARK: - SendStep
@@ -36,8 +40,8 @@ extension OnrampStep: SendStep {
     var sendStepViewAnimatable: any SendStepViewAnimatable { viewModel }
 
     var navigationTrailingViewType: SendStepNavigationTrailingViewType? {
-        .dotsButton { [weak router = viewModel.onrampProvidersCompactViewModel.router] in
-            router?.openOnrampSettingsView()
+        .dotsButton { [weak self] in
+            self?.viewModel.openOnrampSettingsView()
         }
     }
