+++ Tangem/Modules/Send/UI/OnrampWebView/Interactor/OnrampWebViewInteractor.swift
@@ -1,0 +1,7 @@
+//
+//  OnrampWebViewInteractor.swift
+//  TangemApp
+//
+//  Created by Sergey Balashov on 12.11.2024.
+//  Copyright © 2024 Tangem AG. All rights reserved.
+//
