+++ Tangem/Preview Content/Mocks/Send/SendRoutableMock.swift
@@ -26,4 +26,5 @@ class SendRoutableMock: SendRoutable {
     func openOnrampCurrencySelector() {}
     func openOnrampProviders(providersBuilder: OnrampProvidersBuilder, paymentMethodsBuilder: OnrampPaymentMethodsBuilder) {}
     func openOnrampRedirecting(onrampRedirectingBuilder: OnrampRedirectingBuilder) {}
+    func openOnrampWebView(settings: OnrampWebViewViewModel.Settings) {}
 }
